# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Solennix API
  version: 1.0.0
  description: Event management platform for Latin American organizers
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: https://api.solennix.com
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UnavailableDate:
      type: object
      required:
        - id
        - user_id
        - start_date
        - end_date
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
        end_date:
          type: string
        reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UnavailableDateCreateRequest:
      type: object
      required:
        - start_date
        - end_date
      properties:
        start_date:
          type: string
        end_date:
          type: string
        reason:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code or type
        message:
          type: string
          description: Human-readable error message
...
paths:
...
  /api/unavailable-dates:
    get:
      operationId: getUnavailableDates
      summary: Get unavailable dates in a date range
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
        - name: end
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unavailable dates found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnavailableDate"
        "400":
          description: Invalid or missing date parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
    post:
      operationId: createUnavailableDate
      summary: Create an unavailable date range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnavailableDateCreateRequest"
      responses:
        "201":
          description: Unavailable date created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnavailableDate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
  /api/unavailable-dates/{id}:
    delete:
      operationId: deleteUnavailableDate
      summary: Delete an unavailable date range
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Unavailable date deleted
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unavailable date not found
...